Preserve binary compatibility for Database.graph field.
The type of this field has been reverted from GraphDatabaseAPI back to AbstractGraphDatabase
to maintain binary compatibility for code that has already been compiled against the field with
its previous type, as released in Neo4j 1.7.
This commit also changes most direct usage of the field with calls to the getGraph() accessor
which returns the newer GraphDatabaseAPI type, so ultimately dependencies on AbstractGraphDatabase have still been reduced.
It wasn't sensible to replace all usages of the field couldn't be replaced by using the accessor,
in particular Database and its subclasses still access the field directly, and take its type for their constructor parameters.
The field will be made private in Neo4j 1.10 and most occurrence of AbstractGraphDatabase in this area can then be removed.